add BackdoorExploit mod
BackdoorExploit creates a single executable that uses very little RAM and handles both the creation and activation of a "backdoor" on target computers. The program installs a backdoor if you have admin permissions on the target, or attempts to activate it (and silently acquire admin permissions) if you don't. This is NOT treated as a hostile action, and is NOT blocked by a proxy OR a firewall.

For extensions that wish to control the usage of the backdoor, there are two flags that can be set and unset to limit what it's allowed to do, and one that prevents it from being simply given to the user on OS load:
- `NO_INSTALL_BACKDOOR` will disable the ability to install the backdoor exploit on a target computer. Please note that there's nothing stopping the user from connecting to a computer that has it installed, downloading the `/sys/netcfgx.dll` file, connecting to a *different* computer, and then *uploading* that file to effectively "manually" install the exploit.
- `NO_TRIGGER_BACKDOOR` will disable the ability to *use* the backdoor to silently acquire admin permissions. This might be useful to turn it into a mission objective, by allowing it to be installed and then waiting until the target computer has the right contents in the appropriate file.
- `NO_ADD_BACKDOOR_EXE` will prevent the program from being automatically added to the user's computer on OS load.

For extensions wishing to use the backdoor exploit in some way, there are four IDs that will be replaced with the relevant data:
- `#BACKDOOR_EXE#` transforms to the "magic exe data" for the `Backdoor.exe` program
- `#BACKDOOR_EXPLOIT_FOLDER#` transforms to the containing folder for the backdoor exploit file
- `#BACKDOOR_EXPLOIT_FILENAME#` transforms to the name of the file that must contain the magic data
- `#BACKDOOR_EXPLOIT_CONTENT#` transforms to the magic data that must be in the target file
The result is that the file at "`#BACKDOOR_EXPLOIT_FOLDER#`/`#BACKDOOR_EXPLOIT_FILENAME#`" must contain ONLY the EXACT contents `#BACKDOOR_EXPLOIT_CONTENT#` in order for the backdoor to provide admin access. If this is not the case AND the player has admin permissions when they attempt to run the program, it will instead try to INSTALL the backdoor (writing that content to that file).

This will Eventually(tm) be in a readme file.