add Foxnet mod
Foxnet is a work-in-progress development/debugging utility mod, intended to provide ways to manipulate computer nodes. If used outside of development and debugging, it can only be considered a cheat mod. However, nothing is done automatically, so it's fine to leave it installed when you're done.

Documentation will be coming, probably around the time that I refactor the entire command setup.